improved dungeon fighter demo #9

+++ desktop/src/de/ur/mi/examples/dungeonfighter/Vector.java
@@ -33,7 +33,7 @@ public class Vector extends Point {
         return new Vector(getX()-substitution.getX(),getY()-substitution.getY());
     }
 
-    public Vector rotate (float degrees) {
+    public Vector rotate (double degrees) {
         double degreesToRadians = Math.PI / 180;
         double radians = degrees * degreesToRadians;
